Pine: Create new library with strings, math and std types
Pine is meant to be a library shared between kernelspace and userspace
(assuming I get that far in the project). This commit kicks off that
library with a new String type (a fancy wrapper around char*), and a
really basic printf implementation.

Since the printf implementation allowed for numbers in the format, this
also required a atoi implementation among a couple other math functions.
Those math functions aren't efficient nor general, so they are meant to
be temporary.

I'll note that with Pine I'm not trying to recreate libc or libcxx,
though obviously I may replicate significant parts for convience and
practicality sake. Pinyon is a greensfield project with no restrictions
or existing requirements, and I think Pine should be a reflection of
that.