Pine: Introduce new FwdIter template
This allows us to more easily iterate over abstract containers in
the future. Presently, String is the only user of this.

The way a container class would use this is by creating a templated type
of the FwdIter within their class definition and returning the result of
the static begin() and end() methods of the FwdIter in the class'
begin() and end() methods. e.g. for String:

  using ConstIter = FwdIter<const String, const char>;
  ConstIter begin() const { return ConstIter::begin(*this); }
  ConstIter end() const { return ConstIter::end(*this); }

This design keeps the begin()/end() methods private to the FwdIter
templated class, while still reusing generic override operations such
as ++, and the custom at_end() in the FwdIter template.

This design was inspired by SerenityOS's AK/Iterator.h template class,
which IMO is a pretty clean design.