Meta: Track dependencies in GNU-based Makefile to prevent recompile
The downside to this change is that I'm using GNU Makefiles now, rather
than POSIX. In ideal world we'd make portable Makefiles, but they have
been a pain to deal with (for now) and if I pretend that others will run
Pinyon, it shouldn't be that big of a deal to install GNU make along
with GNU GCC.

Also, header file tracking is not currently done. This should be added
in the future.