Kernel: Create simple kmalloc/kfree methods
The way this is implemented is by using doubly-linked headers at the
of each allocation. Finding a free block is done by traversing from the
first block onwards until an available block within size is found. Like
all things in this project, this can vastly be improved in the future,
starting first with schemes to avoid starting at the first block upon
every kmalloc.

Currently we do not have a MMU, so these methods just take what's
available within the middle of our physical address space.