Pine+Kernel: Support %u and %l* in printf; use GCC printf attribute
This commit overhauls printf and allows for format specifiers such as %l,
as well as supporting the %u unsigned integer type.

Part of this change is the move towards using const char* instead of
StringView& in the console functions. Although relying on our custom
StringView class is theoretically safer, it also introduces some overhead
and furthermore, we cannot use GCC's printf attribute with it. This
attribute on consolef permits GCC to start warning about incorrect
formatting, which has already caught a bug!