Kernel: Handle IRQ interrupts; set up system timer
This mega commit _finally_ enables IRQ interrupts and uses the Raspberry
Pi 2's system timer to start to keep track of time.

To demonstrate this I created an "uptime" command in the shell. This
utility reports the time since boot in seconds, though it appears to
drift significantly with real life time, so that will need to be
resolved in the future.