Kernel: Implement basic cooperative context switching
This mega commit basically just makes it possible for a "swi 0" syscall
to cause a new scheduler to switch to running a totally diferent
process.

Among other things, part of this change is breaking out the shell into
it's own task and file. I also got paranoid about device memory, so I
added a MemoryBarrier RAII class; this probably isn't necessary.

To test and demonstrate this, a new command to the shell was added. The
"yield" command will now yield to a spin task that will immediately
yield back to the shell! So useful I know.

It should be noted that this feature is half-baked, if that. Presently,
all tasks still run in supervisor mode and they do not have any isolation
with respect to one another. Furthermore, there is no task cleanup
mechanisms, so if a task exits it's function, the kernel will halt.
Perhaps the most glaring issue though, is the fact that the tasks that
run are hardcoded in the kernel.