Kernel: Create basic sleep syscall and syscall infrastructure
This commit simply adds some basic infrastructure around syscalls. With
this, a new sleep() syscall was added that causes the running task to
sleep for the period in seconds specified. Another task will get
scheduled during that time.

We can ensure that there is always a task that will run with the "spin"
task, which literally just yield()s as soon as scheduled, so it can
always be scheduled. Yielding constantly is not ideal, and this should
be changed in the future, but it is sufficient for now.

The way I've set up syscalls to work is by utilizing the numeric number
found in the swi call to indicate which syscall to use, and utilizing
the r0 register as a possible argument register. This argument is only
applicable to some syscalls. I believe this approach is actually
different from Linux, which utilizes r7 as the syscall number and r0-r6
as the syscall arguments.