Kernel: Obtain console input in shell from syscall; write using syscall
This commit is an incremental change that moves towards a more
independent "userspace". Rather than using console functions directly,
the shell task now asks the kernel for input and asks the kernel to
write data on it's behalf.

UART IO still relies on polling, so currently the scheduler cannot
reschedule things when a task is waiting on input.

Further work also needs to be done to facilitate other shell
functionality such as memory reporting or uptime using syscalls.

To provide the independence of a "userspace", a new lib.hpp library has
been created that will aim to fufill all "userspace" needs using
syscalls.