Pine+Kernel: Ensure dynamic init statics are defined before use
And also place memory barriers after their definition.

When it comes to dynamically initialized statics, like the following
statement:

static TaskManager g_task_manager = TaskManager {};

the point in the program where g_task_manager is initialized with
respect to other statics in other .cpp files actually hard to accurately
say. This means it is possible that g_task_manager is _undefined_ when
you try and use it elsewhere. I actually ran into this issue with that
example and had to do a dirty hack of reinitializating the variable when
I first wanted to use it (see tasks.cpp).

The fix to this problem is to use the "construct on first use" idiom,
which this commit applies.

The following link provides more detail and background on the problem:
https://isocpp.org/wiki/faq/ctors#static-init-order

Buttt.... this is a bit of a can of worms since from testing it looks
like function-scoped static variables that are dynamically initialized
encounter issues with weakly ordered memory (which ARM appears to default
to for memory regions?). What this appears to mean is that the compiler
will usually place a gaurd after your initialiazation. Unfortunately,
for reasons beyond my knowledge (and googling), I get the following
linking issue, probably related to the -ffreestanding nature of the
kernel:

undefined reference to `__cxa_guard_acquire'
undefined reference to `__cxa_guard_release'

So to fix that additional problem, this commit also adds a
'-fno-threadsafe-statics' compilation flag to stop the linker from
complaining, but also places a memory barrier after each initialization.
This barrier is probably overkill in the sense that I think you actually
don't need a full "dmb" instruction, but rather a guard variable. I don't
know enough at this point to confidently do anything but completely
forcing a full-on memory barrier.