Pine+Kernel: Make "userspace" malloc, memory obtained via syscalls
This mega commit creates a per-task malloc based on the [poor]
implementation previously found in kmalloc. The way this is done is by
having a templated MemoryAllocation class in pine/malloc.hpp that both
kmalloc and lib use. The MemoryAllocation class relies on a MemoryBounds
object to reserve additional memory and to tell it where to place memory.

Heap memory for a task is first requested and reserved upon a
heap_reserve() syscall, and the recorded heap memory size is increased
when heap_incr is called. It should be noted that for simplicity sake
each task's heap size is fixed; increasing a task's heap size simply
causes the kernel to note that their heap size is bigger. This is meant
to be temporary until a better scheme, likely using virtual memory can
be constructed (our heap has to be contigious since we're using physical
memory, so dynamically sizing different task's physical memory
allocations is not something I want to be in the business of).