Kernel: Allow for IRQ interrupts during software interrupt handlers
Allowing this effectively means that premption can occur even when
tasks that are waiting on the readline() and write() syscalls in the
software interrupt handler (due to those syscalls still using polling).

This is presently quite hacky, but it appears to work. The solution I
have is hacky in the sense that it relies on implicit assumptions
about the interrupt state of each task after context switching. In
particular, it assumes that interrupts will always be enabled when a
task is running and that all code paths properly ensure that.

The "correct" solution is likely to start saving and modifying the CPSR
register when context switching and entering into the IRQ or SWI
handler.