Kernel+Userspace: Track per-task CPU time and export via syscall
The CPU time of a process can be obtained using the cputime() syscall,
which returns the number of jiffies since boot that the task has spent
executing. The 'uptime' command in the shell was updated to show this.

You will likely notice that the CPU time of the shell is almost exactly
50%. This is because there are only two tasks and the scheduler is round
robben. Sleeping _should_ reduce this ratio slightly.