Kernel: Hack around system timer race conditions
The system timer peripheral on the BCM2835 processor has an interesting
design choice. You must re-initialize the timer match value when you get
a system timer IRQ.

Presumably this is fine on bare metal hardware, but it appears to cause
significant problems in emulation. In particular, if one assumes that
the time required to reinit is always going to be less than the amount
of time you want to wait, then it is possible that you would set the
timer match value to be less than the current timer counter, resulting
in no IRQ being raised. This is because the system timer is accurate
relative to real time, and the emulation of the system is not.

I made this assumption in the kernel and the result was that the current
task would never be preempted sometimes because no IRQs were being
raised. This would either freeze the system if the spin task was
scheduled, or time would stop being counted (and also freeze the system
upon yield()).

Furthermore, in testing I found that I could not just check for whether
the comparison would be less than the current counter and take
appropriate action because even then, race conditions would occur and I
would end up with system freezes. My best guess for this is that because
QEMU is entirely ran in userspace (I believe), it is possible for the
host OS's (MacOS) scheduler to cause the race condition. This is further
validated by the fact that this issue occurs only when I don't enter
input into the shell for a while.

My dirty hack, which appears to work somehow, is to use another
comparison line with a much longer comparison time to ensure an IRQ is
always raised. When I detect that the longer IRQ was triggered, I
increase the jiffies counter accordingly. This period seems to win over
in the race condition always.