Pine: Create a Maybe (std::optional) templated class
There are many cases in the kernel where it is desirable to express that
the return value of a function may not exist. For example, a memory
allocation might fail, and it is preferable in that case to force the
caller to explicitly check for that condition on return, rather than
relying on discipline-based nullptr checks.

This commit recreates the std::optional class found in the stdlib.
Functions that might fail in the kernel and userspace were updated to
use this functionality, though currently the main use case---memory
allocation---still isn't handled properly in the kernel, though the
Maybe return values somewhat make those issues more visually apparent.