Properly release and return movable values in Maybe
I was doing weird things with R-Values when releasing the value; though
I think somehow the code still worked somehow based on the fact that I
have passing tests for these things.

I also used ref-qualifiers [1] to prevent copies when an R-Value is
called for with .value() (or at least a move() is not needed to prevent
a copy).

[1] https://akrzemi1.wordpress.com/2014/06/02/ref-qualifiers/