Simplify ownership of UARTResource
Previously, we had a UARTResource which gave out UARTRequest with a
Owner smart pointer, which had a special destructor to notify
UARTResource that the resource was destroyed/"returned".

I'm not sure having two classes for this, along with a special destructor
was necessary and added any value, so I've changed things so that we now
have a singular UARTResource which tasks request. When that resource is
destructed and the associated memory freed, a static pointer is set to
null to indicate that the resource was returned.