Document interrupt disabling requirements in funcs with parameter tag
In some functions we need to assume that interrupts are disabled. This
fact has been implicit, which can lead to hard to diagnose bugs if we
either didn't disable interrupts before calling the function, or
disabled interrupts again in the function, causing interupts to be
disabled earlier than they should be.

This commit adds a zero-sized verification tag that the programmer must
provide by either decaying the InterruptDisabler to this tag or
promising that interrupts are disabled with a static creation function.
This should hopefully make future mistakes less likely, if this tag is
used wherever interrupts are required to be disabled.