Run tasks in user mode, instead of supervisor
This was a longstanding issue that stems from the fact that I originally
wrote this kernel for a final project in university (little time, lots
of ambition).

This required introducing kernel stacks for each task, and ensuring that
the CPSR is both properly saved in interrupts, as well as saved when
switching between tasks. In addition, the banked SP and LR registers
of user mode must also be saved when switching.

Instead of saving these things on the stack, I now store them in a
struct so that the resuming and starting code can be exactly the same.
It also helps with debugability. If we need to adjust the values of
registers when starting a task, this also makes things easier.

This also fixes an annoying workaround for us clobbering LR registers
when we did syscalls, since LR was clobbered by the vector interrupt but
the "userspace" LR was never automatically saved via banked registers.